+++ src/main/db/index.ts
@@ -2,4 +2,6 @@ export { FirestoreDatabase } from '#/main/db/FirestoreDatabase';
 export { MySqlDatabase } from '#/main/db/MySqlDatabase';
 export { PostgresDatabase } from '#/main/db/PostgresDatabase';
 export { SqliteDatabase } from '#/main/db/SqliteDatabase';
+export { RoutingDatabase } from '#/main/db/RoutingDatabase';
+export { createDatabaseInstance } from '#/main/db/createDatabaseInstance';
 export type { IDatabase } from '#/main/db/IDatabase';